system-plugin: rename main files to plugin.c and plugin.h
